added the final functions but the algorithm is underperforming. Find out why?

+++ lpfgopt/include/lpfgopt.h
@@ -64,7 +64,7 @@ LPFGOPTAPI double* LPFGOPTCALL minimize(
                 double (*fptr)(double*), double* lower, double* upper, 
                 size_t xlen, size_t points, double (*gptr)(double*), 
                 size_t* discrete, size_t discretelen, size_t maxit, 
-                double tol, int seedval);
+                double tol, size_t seedval);
 
 #ifdef __cplusplus
 } // __cplusplus defined.
